Fix rendering calls for imgui.
